Image_app/requirements.txt
transformers
torch
flask
Pillow